chore: enforce consistent typing and formatting across codebase
- Replace float arrays with explicit integer or float types, e.g., `dtype=int32` or `dtype=float32`.
- Improve type safety by adding explicit casting in key computations.
- Update Ruff ignore rules for additional per-file scenarios.
- Refactor imports for cleaner and more direct typing references.

+++ src/supervision/detection/tools/smoother.py
@@ -48,12 +48,12 @@ class DetectionsSmoother:
         ... )
         >>> smoothed = smoother.update_with_detections(detections_1)
         >>> smoothed.xyxy
-        array([[ 0.,  0., 10., 10.]])
+        array([[ 0.,  0., 10., 10.]], dtype=float32)
         >>> smoothed = smoother.update_with_detections(detections_2)
         >>> smoothed.xyxy
-        array([[ 1.,  1., 11., 11.]])
+        array([[ 1.,  1., 11., 11.]], dtype=float32)
         >>> smoothed.confidence
-        array([0.6])
+        array([0.6], dtype=float32)
 
         ```
