+++ src/supervision/metrics/detection.py
@@ -144,8 +144,8 @@ class ConfusionMatrix:
             ...     classes=['person']
             ... )
             >>> confusion_matrix.matrix
-            array([[1., 0.],
-                   [0., 0.]])
+            array([[1, 0],
+                   [0, 0]], dtype=int32)
 
             ```
         """
@@ -218,9 +218,9 @@ class ConfusionMatrix:
             ...     classes=['person', 'dog']
             ... )
             >>> confusion_matrix.matrix
-            array([[1., 0., 1.],
-                   [0., 1., 0.],
-                   [1., 0., 0.]])
+            array([[1, 0, 1],
+                   [0, 1, 0],
+                   [1, 0, 0]], dtype=int32)
 
             ```
         """
